+++ edgelet/support-bundle/src/support_bundle.rs
@@ -6,9 +6,6 @@ use std::fs::File;
 use std::io::{Cursor, Read, Seek, Write};
 use std::path::Path;
 
-// TODO: make tokio
-use std::process::Command as ShellCommand;
-
 use chrono::{DateTime, Local, NaiveDateTime, Utc};
 use failure::Fail;
 use futures::{Future, Stream};
@@ -17,7 +14,7 @@ use zip::{write::FileOptions, CompressionMethod, ZipWriter};
 use edgelet_core::{LogOptions, LogTail, Module, ModuleRuntime};
 
 use crate::error::{Error, ErrorKind};
-use crate::logs::pull_logs;
+use crate::runtime_util::{get_modules, write_check, write_logs};
 
 pub async fn make_bundle<M>(
     output_location: OutputLocation,
@@ -37,11 +34,21 @@ where
     let mut zip_writer = ZipWriter::new(buffer);
 
     // Write all files
-    zip_writer
-        .start_file("check.json", file_options)
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-    write_check(&mut zip_writer, iothub_hostname).await?;
+    {
+        // Get Check
+        zip_writer
+            .start_file("check.json", file_options)
+            .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
+        write_check(&mut zip_writer, iothub_hostname).await?;
 
+        for module_name in get_modules(&runtime, include_ms_only).await? {
+            // Get all logs
+            zip_writer
+                .start_file(format!("logs/{}_log.txt", module_name), file_options)
+                .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
+            write_logs(&runtime, &module_name, &log_options, &mut zip_writer).await?;
+        }
+    }
     // Finilize buffer and set cursur to 0 for reading.
     let mut buffer = zip_writer
         .finish()
@@ -57,46 +64,4 @@ where
 pub enum OutputLocation {
     File(OsString),
     Memory,
-}
-
-async fn write_check(
-    writer: &mut impl Write,
-    iothub_hostname: Option<String>,
-) -> Result<(), Error> {
-    print_verbose("Calling iotedge check");
-
-    let mut iotedge = env::args().next().unwrap();
-    if iotedge.contains("aziot-edged") {
-        print_verbose("Calling iotedge check from edgelet, using iotedge from path");
-        iotedge = "iotedge".to_string();
-    }
-
-    let mut check = ShellCommand::new(iotedge);
-    check.arg("check").args(&["-o", "json"]);
-
-    if let Some(host_name) = iothub_hostname {
-        check.args(&["--iothub-hostname", &host_name]);
-    }
-    let check = check
-        .output()
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-
-    writer
-        .write_all(&check.stdout)
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-    writer
-        .write_all(&check.stderr)
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-
-    print_verbose("Wrote check output to file");
-    Ok(())
-}
-
-fn print_verbose<S>(message: S)
-where
-    S: std::fmt::Display,
-{
-    if true {
-        println!("{}", message);
-    }
 }
